feat: restructure UserInfoResponse DTO to record and update response format
- Move UserInfoResponse from member/dto to member/application/dto/response
- Convert from class with @Builder to record pattern
- Change isAdmin boolean field to role string for better extensibility
- Update UserInfoService to use constructor instead of builder pattern
- Update import statements in UserInfoController and UserInfoService
