--- src/main/java/com/example/nomodel/member/dto/UserInfoResponse.java
@@ -1,36 +1,0 @@
-package com.example.nomodel.member.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@Schema(description = "사용자 정보 응답")
-public class UserInfoResponse {
-
-    @Schema(description = "사용자 고유 ID", example = "1")
-    private Long id;
-
-    @Schema(description = "사용자 이름", example = "홍길동")
-    private String name;
-
-    @Schema(description = "이메일", example = "user@example.com")
-    private String email;
-
-    @Schema(description = "가입일", example = "2024-01-15T10:30:00")
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
-    private LocalDateTime joinedAt;
-
-    @Schema(description = "요금제", allowableValues = {"free", "pro", "enterprise"}, example = "free")
-    private String planType;
-
-    @Schema(description = "현재 포인트", example = "1000")
-    private Integer points;
-
-    @Schema(description = "관리자 여부", example = "false")
-    private Boolean isAdmin;
-}
